requirements.txt
psycopg2-binary~=2.9.11
duckdb~=1.4.4
pandas~=2.3.3
numpy~=2.2.6
matplotlib~=3.10.8
joblib~=1.5.3
scikit-learn~=1.7.2
tensorflow~=2.20.0
SQLAlchemy~=2.0.46
workalendar~=17.0.0